fix(chart): improve time precision and visual consistency in roasting chart
This commit refactors the roasting chart and data tracking to use floating-point interval numbers, enabling more precise plotting of events (like First Crack or Turn Point) that occur between fixed time intervals. It also updates time formatting and UI constraints for better readability.

### Key Changes:

**Data Model & Logic:**
- **Precision Improvements:** Updated `ChartDataPoint` and related logic in `RoastingRunViewModel` and `RoastingDetailScreen` to use `Float` for `intervalNumber`. This allows events to be plotted at exact second offsets rather than rounding to the nearest interval.
- **Enhanced Calculation:** Refactored `totalSeconds` calculation to use `roundToInt()` for better accuracy when converting from float intervals.

**Chart & UI Enhancements:**
- **Refined X-Axis:** Increased axis granularity to `0.1f` and optimized label counts to prevent overlapping text.
- **Time Formatting:** Unified time display to use a `M:SS` format (e.g., "1:30") across the chart and PDF exports, replacing the previous decimal-based or minute-only formats.
- **Interaction Tweaks:** Disabled horizontal and vertical highlight indicators (yellow lines) when selecting points on the chart for a cleaner look.
- **Data Consistency:** Removed manual filtering in the ROR and parameter rows to ensure all data points are rendered consistently across the scrollable area.

**PDF Export:**
- **Formatting Fixes:** Updated `PdfExportManager` to reflect the new time formatting and ensure ROR values are exported as integers.
- **Chart Parity:** Synchronized the PDF chart's axis logic and granularity with the on-screen UI component.

+++ app/src/main/java/com/indie/roastlog/ui/screens/detail/RoastingDetailScreen.kt
@@ -77,7 +77,6 @@ fun RoastingDetail(
                     val interval = s.intervalSeconds
                     val dataMap = s.temperatureData.associateBy { it.intervalNumber }
                     
-                    // Filter out duplicate or inconsistent intervals if any, and ensure sorted
                     val maxIntervalNum = s.temperatureData.maxOfOrNull { it.intervalNumber } ?: 0
                     
                     val points = (0..maxIntervalNum).map { intervalNum ->
@@ -87,11 +86,10 @@ fun RoastingDetail(
                         val prevTemp = if (intervalNum > 0) dataMap[intervalNum - 1]?.temperature else null
                         val rorValue = if (intervalNum > 0 && currentTemp != null && prevTemp != null) currentTemp - prevTemp else null
                         
-                        // Find burner power for this specific time
                         val burnerPower = s.burnerPlan.find { it.seconds == secondsAtThisInterval }?.temperature?.toInt()?.toString() ?: ""
 
                         ChartDataPoint(
-                            intervalNumber = intervalNum,
+                            intervalNumber = intervalNum.toFloat(),
                             totalSeconds = secondsAtThisInterval,
                             temperature = currentTemp,
                             ror = rorValue,
@@ -101,7 +99,7 @@ fun RoastingDetail(
                     
                     listOfNotNull(s.turnPoint, s.yellowing, s.firstCrack, s.endRoast).forEach { ev ->
                         points.add(ChartDataPoint(
-                            intervalNumber = ev.seconds / interval,
+                            intervalNumber = ev.seconds.toFloat() / interval,
                             totalSeconds = ev.seconds,
                             temperature = ev.temperature,
                             ror = null
